+++ src/components/Posts/comments/SingleComment/index.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import db, { auth } from "../../../../firebase";
 import { Button } from "@material-ui/core";
-import { CommentMenuButton } from "../../CommentMenuButton";
+import { CommentMenuButton } from "../CommentMenu/CommentMenuButton";
 import { splitString, isUrl } from "../../urlIdentifier";
 import Container from "./styledSingleComment";
